Skip exporting existing files
If a file already exists, don't bother exporting it. Re-running the
export all (in the case of a crash, error, or cancellation) is
significantly faster, although the script will still go through the long
process of opening every file to check.

This adds the slight risk of missing a file if one is somehow partially
exported, but it does allow "re-starting" an export. Since the export
has a different filename if a new version of anything is saved, this
should support re-exporting everything to capture new+updated designs.