Minor tweaks to the readme, to add a bit more info
Adds in a note about the fact this script also exports an f3d/f3z file
for each file in fusion (that can be exported as one).
Also adds in a "download here" link in the instructions, instead of
just telling people to clone the repo - which should make this tool
appear friendlier to those not familiar with git.